Remove submodule and change to vendor for libaudit-go Changed path for json files required by libaudit-go
Reason: Standard way of installing go programs is by "go get" and our way of
using "./libaudit.go" wouldn't work with "go get" and also it wasn't a
recommended way to structure go packages. One way or another we would have
to revert to standard method. Since we were using libaudit as a submodule
it would now make sense to add it under "vendor/" as many go-packages are
doing today. We can continue to work in similar work fashion as we used to
have under git submodule and can update libaudit-go to upstream only when
we are ready.